requirements.txt
bcrypt==4.0.1
blinker==1.7.0
certifi==2023.7.22
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-Security-Too==5.3.2
Flask-WTF==1.2.1
idna==3.4
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
passlib==1.7.4
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.6.3
urllib3==2.0.2
wcwidth==0.2.6
Werkzeug==3.0.1
WTForms==3.1.1
